# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: SommOS API
  version: 1.0.0
  description: |
    Canonical contract for the SommOS yacht wine management platform. All
    endpoints are prefixed with `/api` and return JSON payloads.
servers:
  - url: /api
    description: Internal Express server base path
  - url: /api/v1
    description: Version 1 API base path
components:
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Stable machine-readable error identifier.
        message:
          type: string
          description: Human friendly explanation of the error.
        details:
          description: Optional structured error context.
          oneOf:
            - type: string
            - type: object
            - type: array
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          const: false
        error:
          $ref: '#/components/schemas/Error'
...
  securitySchemes:
    accessTokenCookie:
      type: apiKey
      in: cookie
      name: sommos_access_token
      description: >-
        Short-lived JWT issued after authentication. Required for protected mutations.
    refreshTokenCookie:
      type: apiKey
      in: cookie
      name: sommos_refresh_token
      description: >-
        Long-lived refresh token stored as an HttpOnly cookie. Used only with the refresh endpoint.
paths:
...
  /system/spec:
    get:
      tags: [System]
      summary: Retrieve the SommOS OpenAPI specification.
      responses:
        '200':
          description: OpenAPI document served as YAML.
          content:
            application/yaml:
              schema:
                type: string
        '500':
          description: Specification is unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'